requirements.txt
langchain==0.1.12
openai==1.14.1
pandas==2.0.3
Pillow==9.5.0
Pillow==10.2.0
python_docx==1.1.0
streamlit==1.26.0
XlsxWriter==3.1.9
